% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_pvalues.R
\name{categorize_pvalues}
\alias{categorize_pvalues}
\title{Categorize p-values for beta coefficients in printing the summary() of a linreg object}
\usage{
categorize_pvalues(p_values)
}
\arguments{
\item{linregobject}{An object of class linreg}
}
\description{
Internal use in the summary.linreg() function
}
\keyword{internal}
