% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jpg.R
\name{get_jpg}
\alias{get_jpg}
\title{A function for getting a jpg file for adding the logo to plot() of a linreg object}
\usage{
get_jpg(filename)
}
\arguments{
\item{filename}{The name of a png file}
}
\description{
Internal use in the plot.linreg() function
}
\keyword{internal}
