% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Compute multiple regression using QR decomposition}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{A model formula}

\item{data}{A dataframe}
}
\value{
A linreg object
}
\description{
Compute multiple regression using QR decomposition
}
\examples{
fit <- linreg(Petal.Length ~ Sepal.Width + Sepal.Length, data = iris)
}
