% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liu_theme.R
\name{liu_theme}
\alias{liu_theme}
\title{A custom theme for Linkoping University used in plot() of a linreg object}
\usage{
liu_theme()
}
\description{
Internal use in the plot.linreg() function
}
\keyword{internal}
