% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{predict.linreg}
\alias{predict.linreg}
\title{Get fitted values from a linreg object}
\usage{
\method{predict}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class linreg}

\item{...}{Additional arguments that we don't use}
}
\description{
Get fitted values from a linreg object
}
\examples{
fit <- linreg(Petal.Length ~ Sepal.Width + Sepal.Length, data = iris)
predict(fit)
}
