% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{ridgereg}
\alias{ridgereg}
\title{Compute Ridge Regression}
\usage{
ridgereg(formula, data, lambda, use_QR = FALSE)
}
\arguments{
\item{formula}{A model formula}

\item{data}{A dataframe}

\item{lambda}{A numeric scalar}

\item{use_QR}{Boolean for if QR-decomposition should be used}
}
\value{
A ridgereg object
}
\description{
Compute Ridge Regression
}
\details{
This function computes ridge regression using least squares or QR-decomposition.
}
\examples{
fit <- ridgereg(Petal.Length ~ Sepal.Width + Sepal.Length, data = iris, lambda = 2)
fit2 <- ridgereg(Petal.Length ~ Sepal.Width + Sepal.Length, data = iris, lambda = 2, use_QR = TRUE)
}
